Conflicts:
	arch/arm/boot/dts/qcom/OP-batterydata-3300mah.dtsi
	arch/arm/boot/dts/qcom/cheeseburger-dvt1.dtsi
	arch/arm/boot/dts/qcom/cheeseburger-pvt.dtsi
	arch/arm/boot/dts/qcom/cheeseburger-pvt1.dtsi
	arch/arm/boot/dts/qcom/cheeseburger-v2.1-dvt1.dts
	arch/arm/boot/dts/qcom/cheeseburger-v2.1-pvt.dts
	arch/arm/boot/dts/qcom/cheeseburger-v2.1-pvt1.dts
	arch/arm/boot/dts/qcom/cheeseburger.dtsi
	arch/arm/boot/dts/qcom/dsi-panel-samsung_s6e3fa3_1080p_cmd.dtsi
	arch/arm/boot/dts/qcom/dsi-panel-samsung_s6e3fa3_1080p_video.dtsi
	arch/arm/boot/dts/qcom/dsi-panel-samsung_s6e3fa5_1080p_cmd.dtsi
	arch/arm/boot/dts/qcom/dsi-panel-samsung_s6e3fa6_1080p_cmd.dtsi
	arch/arm/boot/dts/qcom/dsi-panel-samsung_s6e3fc1_cmd.dtsi
	arch/arm/boot/dts/qcom/dumpling-dvt.dtsi
	arch/arm/boot/dts/qcom/dumpling-pvt.dtsi
	arch/arm/boot/dts/qcom/dumpling-v2.1-Second_Resource.dts
	arch/arm/boot/dts/qcom/dumpling-v2.1-dvt.dts
	arch/arm/boot/dts/qcom/dumpling-v2.1-pvt.dts
	arch/arm/boot/dts/qcom/dumpling.dtsi
	arch/arm64/include/asm/io.h
	drivers/cpufreq/qcom-cpufreq.c
	drivers/cpuidle/lpm-levels.c
	drivers/devfreq/devfreq_devbw.c
	drivers/input/fingerprint/fpc/fpc1020_tee.c
	drivers/input/fingerprint/goodix/gf_spi.c
	drivers/input/fingerprint/goodix/gf_spi.h
	drivers/input/fingerprint/goodix/platform.c
	drivers/input/misc/Kconfig
	drivers/input/misc/Makefile
	drivers/input/misc/qpnp-power-on.c
	drivers/input/misc/tri_state_key.c
	drivers/input/touchscreen/Kconfig
	drivers/input/touchscreen/fw_update_v7.if
	drivers/input/touchscreen/synaptics_baseline.h
	drivers/input/touchscreen/synaptics_driver_s3320.c
	drivers/input/touchscreen/synaptics_redremote.h
	drivers/input/touchscreen/synaptics_s3320_redremote.c
	drivers/iommu/msm_dma_iommu_mapping.c
	drivers/leds/leds-qpnp.c
	drivers/media/platform/msm/camera_v2/isp/msm_isp_axi_util.c
	drivers/media/platform/msm/camera_v2/sensor/msm_sensor.c
	drivers/media/platform/msm/camera_v2/sensor/msm_sensor_driver.c
	drivers/misc/Makefile
	drivers/nfc/Kconfig
	drivers/nfc/Makefile
        drivers/nfc/pn5xx.c
	drivers/pci/host/pci-msm.c
	drivers/pci/pci.c
	drivers/pinctrl/qcom/pinctrl-msm.c
	drivers/power/power_supply_sysfs.c
	drivers/power/supply/qcom/qpnp-fg-gen3.c
	drivers/scsi/ufs/ufs.h
	drivers/soc/qcom/icnss.c
	drivers/soc/qcom/project_info.c
	drivers/soc/qcom/socinfo.c
	drivers/usb/core/driver.c
        drivers/usb/core/hub.c
	drivers/usb/dwc3/core.c
	drivers/usb/dwc3/ep0.c
	drivers/usb/dwc3/gadget.c
	drivers/usb/gadget/composite.c
	drivers/usb/pd/policy_engine.c
	drivers/video/fbdev/core/fbmem.c
	drivers/video/fbdev/msm/Kconfig
	drivers/video/fbdev/msm/mdss_dsi_panel.c
	fs/sdcardfs/inode.c
	include/linux/cpufreq.h
	include/linux/gpio/driver.h
	include/linux/input/qpnp-power-on.h
	include/linux/pm_qos.h
	include/linux/power/oem_external_fg.h
	include/linux/power_supply.h
	include/linux/project_info.h
	include/soc/qcom/smem.h
	kernel/power/process.c
	kernel/power/qos.c
	kernel/power/suspend.c
	net/wireless/db.txt
	sound/soc/codecs/Makefile
	sound/soc/codecs/tfa9891/tfa9897_init.c
	sound/soc/codecs/tfa9891/tfa98xx.c
	sound/soc/codecs/tfa9891/tfa_dsp.c
	sound/soc/codecs/wcd-mbhc-v2.c
	sound/soc/codecs/wcd-mbhc-v2.h
	sound/soc/codecs/wcd9335.c
	sound/soc/codecs/wcd9335.h
	sound/soc/msm/msm8998.c
	sound/soc/msm/qdsp6v2/msm-pcm-routing-v2.c

+++ sound/soc/codecs/tfa9891/tfa_dsp.c
@@ -24,8 +24,6 @@
 #include "tfa_dsp_fw.h"
 /* TODO: remove genregs usage? */
 #include "tfa98xx_genregs_N1C.h"
-/*zhiguang.su@MultiMedia.AudioDrv, 2015-11-09, add for debug*/
-#include <sound/sounddebug.h>
 /* handle macro for bitfield */
 #define TFA_MK_BF(reg, pos, len) ((reg<<8)|(pos<<4)|(len-1))
 
@@ -2562,7 +2560,7 @@ enum Tfa98xx_Error tfaRunSpeakerCalibration(Tfa98xx_handle_t handle, int profile
 	enum Tfa98xx_Error err = Tfa98xx_Error_Ok;
 	int calibrateDone, spkr_count = 0;
 	/* Avoid warning in user-space */
-	profile = (int)profile;
+	profile=profile;
 
 #ifdef __KERNEL__ /* Necessary otherwise we are thrown out of operating mode in kernel (because of internal clock) */
 	if((strstr(tfaContProfileName(handle, profile), ".cal") == NULL) && (tfa98xx_dev_family(handle) == 2))
